Updated face data export to include the 1st UV set
Was hoping that the redundant UV data in the face data section is unnecessary since the introduction of the #UVSet# TAGG sections. That turned out to not be the case...
The game falls back to this data for color textures, and when the UV index is not set explicitly in the rvmat.