Simplified normal processing
Face and loop indices are in the same order in the blender data as they are read from the P3D apparently. This means that the split normal list can be constructed by simply iterating over the face data, no need to query the mesh itself. This makes it simpler, and faster.