Small bug fix in voronoi cell generation
The search radius must be twice the current bounding sphere radius (as the actual cutting plane is halfway, so it's the halfway point that needs to be inside the radius). Since the shape quickly converges to the actual shape with the distance ordered vertex processing, the error with single radius is usually not very large, but an error nonetheless.